refactor(frontend): reduce file complexity in DiceRoller.tsx by extracting logic into sub-modules
